Modificacion en migration

+++ src/database/migrations/2020_05_06_000001_create_comprobantes_table.php
@@ -18,13 +18,13 @@ class CreateComprobantesTable extends Migration
             $table->integer('tipo');
             $table->integer('numero');
             $table->integer('cliente_tipo_doc')->nullable();
-            $table->integer('cliente_num_doc')->nullable();
+            $table->unsignedBigInteger('cliente_num_doc')->nullable();
             $table->date('fecha');
             $table->float('importe_total',10,2);
             $table->float('importe_neto',10,2);
             $table->float('importe_iva',10,2);
             $table->integer('comprobante_asociado')->nullable();
-            $table->bigInteger('cae')->nullable();
+            $table->unsignedBigInteger('cae')->nullable();
             $table->date('vencimiento_cae')->nullable();
             $table->string('resultado',1);
             $table->timestamps();
